Shuffle around classes and files
Files and classes have been reorganized to be more maintainable.
