+++ include/fastnumbers/buffer.hpp
