+++ include/fastnumbers/parser.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "fastnumbers/parser/base.hpp"
-#include "fastnumbers/parser/buffer.hpp"
 #include "fastnumbers/parser/character.hpp"
 #include "fastnumbers/parser/numeric.hpp"
 #include "fastnumbers/parser/unicode.hpp"
