+++ include/fastnumbers/parser/unicode.hpp
@@ -2,7 +2,6 @@
 
 #include <Python.h>
 
-#include "fastnumbers/EnumClass.h"
 #include "fastnumbers/parser/base.hpp"
 #include "fastnumbers/user_options.hpp"
